backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
scipy>=1.11.3
llama-index-core>=0.10.0
llama-index-vector-stores-faiss>=0.3.0
llama-index-embeddings-siliconflow>=0.2.1
llama-index-llms-siliconflow>=0.1.0
llama-index-embeddings-openai>=0.3.0
faiss-cpu>=1.7.4
pickle5>=0.0.11